+++ .storybook/blocks/grid/Col.tsx
@@ -1,4 +1,4 @@
-import React, { InputHTMLAttributes, useCallback, ChangeEvent, ReactNode } from 'react';
+import React, { ReactNode } from 'react';
 import cn from 'classnames';
 
 import styles from './Col.module.css';
@@ -10,13 +10,5 @@ type ColProps = {
 };
 
 export const Col: React.FC<ColProps> = ({ children, className, span = 12 }) => (
-    <div
-        className={cn(
-            className,
-            styles.col,
-            styles[`span-${span}`],
-        )}
-    >
-        {children}
-    </div>
+    <div className={cn(className, styles.col, styles[`span-${span}`])}>{children}</div>
 );