+++ packages/cdn-icon/src/Component.tsx
@@ -17,17 +17,28 @@ type CDNIconProps = {
      */
     className?: string;
     /**
+     * Базовый адрес cdn хранилища c иконками
+     * @default https://alfabank.servicecdn.ru/icons
+     */
+    baseUrl?: string;
+    /**
      * Идентификатор для систем автоматизированного тестирования
      */
     dataTestId?: string;
 };
 
-export const CDNIcon: React.FC<CDNIconProps> = ({ name, color, dataTestId, className }) => {
+export const CDNIcon: React.FC<CDNIconProps> = ({
+    name,
+    color,
+    dataTestId,
+    className,
+    baseUrl = 'https://alfabank.servicecdn.ru/icons',
+}) => {
     const [icon, setIcon] = useState('');
 
     useEffect(() => {
         const xhr = new XMLHttpRequest();
-        xhr.open('GET', `https://alfabank.gcdn.co/icons/${name}.svg`);
+        xhr.open('GET', `${baseUrl}/${name}.svg`);
         xhr.send();
         xhr.onload = function onload() {
             const svg = xhr.response;
@@ -35,7 +46,7 @@ export const CDNIcon: React.FC<CDNIconProps> = ({ name, color, dataTestId, class
         };
 
         return () => xhr.abort();
-    }, [name]);
+    }, [name, baseUrl]);
 
     return (
         <span