+++ packages/plate/src/Component.tsx
@@ -186,17 +186,19 @@ export const Plate = forwardRef<HTMLDivElement, PlateProps>(
         const hasAnyAddons = leftAddons || subAddons || foldable || hasCloser;
 
         const handleClick = useCallback(
-            event => {
-                const eventInsideComponent =
-                    plateRef.current && plateRef.current.contains(event.target);
+            (event: React.MouseEvent<HTMLDivElement> | React.KeyboardEvent<HTMLDivElement>) => {
+                const target = event.target as HTMLDivElement;
+                const eventInsideComponent = plateRef.current && plateRef.current.contains(target);
 
                 const eventInsideContent =
-                    contentRef.current && contentRef.current.contains(event.target);
+                    contentRef.current && contentRef.current.contains(target);
 
                 const eventInsideSubAddons =
                     subAddonsRef.current && subAddonsRef.current.contains(event.target);
-
-                const clickSimilarKeys = ['Enter', ' '].includes(event.key);
+                
+                const clickSimilarKeys = ['Enter', ' '].includes(
+                        (event as React.KeyboardEvent<HTMLDivElement>).key,
+                    );
 
                 const shouldChangeIsFolded =
                     eventInsideComponent &&
@@ -215,14 +217,14 @@ export const Plate = forwardRef<HTMLDivElement, PlateProps>(
                 }
 
                 if (onClick) {
-                    onClick(event);
+                    onClick(event as React.MouseEvent<HTMLDivElement>);
                 }
             },
             [foldable, onClick, uncontrolled, onToggle, foldedState, foldedProp],
         );
 
         const handleClose = useCallback(
-            event => {
+            (event: React.MouseEvent<HTMLButtonElement>) => {
                 setIsHidden(true);
 
                 if (onClose) {