+++ packages/calendar/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.6.0](https://github.com/core-ds/core-components/compare/@alfalab/core-components-calendar@5.5.8...@alfalab/core-components-calendar@5.6.0) (2022-08-04)
+
+
+### Features
+
+* react 18 support ([#159](https://github.com/core-ds/core-components/issues/159)) ([2e6693c](https://github.com/core-ds/core-components/commit/2e6693c62f534e333aadb7d3fff4ffd78ac84c63))
+
+
+
+
+
+## [5.5.8](https://github.com/core-ds/core-components/compare/@alfalab/core-components-calendar@5.5.7...@alfalab/core-components-calendar@5.5.8) (2022-07-25)
+
+**Note:** Version bump only for package @alfalab/core-components-calendar
+
+
+
+
+
 ## [5.5.7](https://github.com/core-ds/core-components/compare/@alfalab/core-components-calendar@5.5.6...@alfalab/core-components-calendar@5.5.7) (2022-07-18)
 
 **Note:** Version bump only for package @alfalab/core-components-calendar