+++ packages/slider/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.1.0](https://github.com/core-ds/core-components/compare/@alfalab/core-components-slider@3.0.4...@alfalab/core-components-slider@3.1.0) (2022-08-04)
+
+
+### Bug Fixes
+
+* **slider:** заменил interface на type ([#176](https://github.com/core-ds/core-components/issues/176)) ([d19e3cb](https://github.com/core-ds/core-components/commit/d19e3cb3a728feb6a0dc46a6a0691f72fc90e10f))
+
+
+### Features
+
+* react 18 support ([#159](https://github.com/core-ds/core-components/issues/159)) ([2e6693c](https://github.com/core-ds/core-components/commit/2e6693c62f534e333aadb7d3fff4ffd78ac84c63))
+
+
+
+
+
 ## [3.0.4](https://github.com/core-ds/core-components/compare/@alfalab/core-components-slider@3.0.3...@alfalab/core-components-slider@3.0.4) (2022-07-18)
 
 **Note:** Version bump only for package @alfalab/core-components-slider